+++ src/esm/utils/detectDarkMode.js
@@ -8,21 +8,31 @@
  * @returns {boolean}
  */
 export const detectDarkMode = (GUP) => {
+    console.log("detectDarkMode", detectDarkMode);
+    var isDarkmode = false;
     // Respect overrides
     if (typeof GUP !== "undefined" && GUP.hasOwnProperty("darkmode") && GUP["darkmode"]) {
         // const overrideValue = PlotBoilerplate.utils.fetch.bool(GUP, "darkmode", null);
         const overrideValue = !!JSON.parse(GUP["darkmode"]);
         if (overrideValue !== null) {
-            return overrideValue;
+            isDarkmode = overrideValue;
         }
     }
-    if (globalThis.matchMedia && globalThis.matchMedia("(prefers-color-scheme: dark)").matches) {
+    else if (globalThis.matchMedia && globalThis.matchMedia("(prefers-color-scheme: dark)").matches) {
         // dark mode by system
-        return true;
+        isDarkmode = true;
     }
-    // else: dark mode by daytime
-    const hours = new Date().getHours();
-    const isDayTime = hours > 6 && hours < 18;
-    return !isDayTime;
+    else {
+        // else: dark mode by daytime
+        const hours = new Date().getHours();
+        const isDayTime = hours > 6 && hours < 18;
+        isDarkmode = !isDayTime;
+    }
+    console.log("Adding darkmode to <body>", isDarkmode);
+    if (isDarkmode) {
+        // Add darkmode class to <body> tag
+        document.body.classList.add("darkmode");
+    }
+    return isDarkmode;
 };
 //# sourceMappingURL=detectDarkMode.js.map