+++ src/ts/utils/detectDarkMode.ts
@@ -9,20 +9,30 @@
  */
 
 export const detectDarkMode = (GUP?: Record<string, string>): boolean => {
+  console.log("detectDarkMode", detectDarkMode);
+  var isDarkmode: boolean = false;
   // Respect overrides
   if (typeof GUP !== "undefined" && GUP.hasOwnProperty("darkmode") && GUP["darkmode"]) {
     // const overrideValue = PlotBoilerplate.utils.fetch.bool(GUP, "darkmode", null);
     const overrideValue = !!JSON.parse(GUP["darkmode"]);
     if (overrideValue !== null) {
-      return overrideValue;
+      isDarkmode = overrideValue;
     }
-  }
-  if (globalThis.matchMedia && globalThis.matchMedia("(prefers-color-scheme: dark)").matches) {
+  } else if (globalThis.matchMedia && globalThis.matchMedia("(prefers-color-scheme: dark)").matches) {
     // dark mode by system
-    return true;
+    isDarkmode = true;
+  } else {
+    // else: dark mode by daytime
+    const hours = new Date().getHours();
+    const isDayTime = hours > 6 && hours < 18;
+    isDarkmode = !isDayTime;
   }
-  // else: dark mode by daytime
-  const hours = new Date().getHours();
-  const isDayTime = hours > 6 && hours < 18;
-  return !isDayTime;
+
+  console.log("Adding darkmode to <body>", isDarkmode);
+  if (isDarkmode) {
+    // Add darkmode class to <body> tag
+    document.body.classList.add("darkmode");
+  }
+
+  return isDarkmode;
 };